requirements.txt
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.1
blessed==1.25.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
comm==0.2.3
confection==0.1.5
contexttimer==0.3.3
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
debugpy==1.8.18
decorator==5.2.1
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
executing==2.2.1
filelock==3.20.2
findspark==2.0.1
fonttools==4.61.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
graphframes==0.6
humanfriendly==10.0
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
linalg==1.0.4
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
murmurhash==1.0.15
mvn==0.1.3
nest-asyncio==1.6.0
networkx==3.6.1
nose==1.3.7
numba==0.63.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
preshed==3.0.12
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.5
pyspark==4.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
stack-data==0.6.3
sympy==1.14.0
thinc==8.3.10
threadpoolctl==3.6.0
torch==2.9.1
torchvision==0.24.1
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.9.post2
urllib3==2.6.2
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
wrapt==2.0.1
